+++ tests/functional/adapter/python_model/fixtures.py
@@ -89,6 +89,7 @@ import pandas
 def model(dbt, spark):
     dbt.config(
         materialized='table',
+        unique_tmp_table_suffix=True
     )
     data = [[1,2]] * 10
     return spark.createDataFrame(data, schema=['test1', 'test3'])
@@ -189,4 +190,27 @@ expected_complex = """date,name
 2,"Teo"
 2,"Fang"
 3,"Elbert"
+"""
+
+simple_incremental_python_model = """
+import pandas
+
+def model(dbt, spark):
+    dbt.config(
+        materialized='incremental',
+    )
+    data = [[1,2]] * 5
+    return spark.createDataFrame(data, schema=['test', 'test2'])
+"""
+
+simple_incremental_python_model_v2 = """
+import pandas
+
+def model(dbt, spark):
+    dbt.config(
+        materialized='incremental',
+        unique_tmp_table_suffix=True,
+    )
+    data = [[1,2]] * 10
+    return spark.createDataFrame(data, schema=['test', 'test2'])
 """